feat : 로또 결과를 책임지는 클래스
