+++ src/main/java/lotto/step2/domain/WinningNumber.java
@@ -11,20 +11,7 @@ public class WinningNumber {
         this.winningNumbers = winningNumbers;
     }
 
-    public void result(ArrayList<ArrayList> lotto){
-        Collections.sort(winningNumbers);
-        int lottoTickets = lotto.size();
-
-        for (int i=0; i<lottoTickets; i++) {
-            Collections.sort(lotto.get(i));
-            int count =0;
-            for (int j=0; j<lotto.get(i).size(); j++) {
-                if (winningNumbers.get(j)== lotto.get(i).get(j)){
-                    count++;
-                }
-            }
-
-
-        }
+    public List<Integer> getWinningNumbers() {
+        return winningNumbers;
     }
 }
